backend/parsers/docx_parser.py
from docx import Document
import io

def extract_text_from_docx(file_bytes: bytes) -> str:
    """Extract text from a DOCX file using python-docx."""
    file_like = io.BytesIO(file_bytes)
    doc = Document(file_like)
    full_text = []
    for para in doc.paragraphs:
        full_text.append(para.text)
    return "\n".join(full_text)
