backend/requirements.txt
fastapi
uvicorn
pdfminer.six
python-docx
pydantic
python-multipart
# No external deps for re as it is built-in.
